+ Restore hooks for tileDrawing + Porting ExampleAnimatedTile to 1.4 * fix ExampleTorch
