+++ ExampleMod/Content/Tiles/ExampleTorch.cs
@@ -137,7 +137,7 @@ namespace ExampleMod.Content.Tiles
 				float xx = Utils.RandomInt(ref randSeed, -10, 11) * 0.15f;
 				float yy = Utils.RandomInt(ref randSeed, -10, 1) * 0.35f;
 
-				spriteBatch.Draw(flameTexture.Value, new Vector2(xx * 16 - (int)Main.screenPosition.X - (width - 16f) / 2f + xx, yy * 16 - (int)Main.screenPosition.Y + offsetY + yy) + zero, new Rectangle(frameX, frameY, width, height), color, 0f, default, 1f, SpriteEffects.None, 0f);
+				spriteBatch.Draw(flameTexture.Value, new Vector2(i * 16 - (int)Main.screenPosition.X - (width - 16f) / 2f + xx, j * 16 - (int)Main.screenPosition.Y + offsetY + yy) + zero, new Rectangle(frameX, frameY, width, height), color, 0f, default, 1f, SpriteEffects.None, 0f);
 			}
 		}
 	}
